Fix issue with autogen description
